+++ composeApp/src/commonMain/kotlin/domain/therapist/TherapistRequestContract.kt
@@ -4,7 +4,7 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GetReviewsRequest(@SerialName("therapist_id") val therapistId: Int)
+data class GetReviewsRequest(@SerialName("therapist_id") val therapistId: Long)
 
 @Serializable
 data class ArchiveAppointmentRequest(@SerialName("appointmentId") val appointmentId: Long)
